Cache worlds for Spigot and Paper
