+++ bukkit/src/main/java/ch/andre601/advancedserverlist/paper/events/PaperEventWrapper.java
@@ -29,7 +29,7 @@ import ch.andre601.advancedserverlist.api.events.GenericServerListEvent;
 import ch.andre601.advancedserverlist.api.objects.GenericServer;
 import ch.andre601.advancedserverlist.api.profiles.ProfileEntry;
 import ch.andre601.advancedserverlist.bukkit.BukkitCore;
-import ch.andre601.advancedserverlist.bukkit.events.PreServerListSetEventImpl;
+import ch.andre601.advancedserverlist.bukkit.listeners.PreServerListSetEventImpl;
 import ch.andre601.advancedserverlist.bukkit.objects.BukkitServerImpl;
 import ch.andre601.advancedserverlist.bukkit.objects.SpigotPlayerImpl;
 import ch.andre601.advancedserverlist.core.interfaces.events.GenericEventWrapper;
@@ -43,7 +43,6 @@ import me.clip.placeholderapi.PlaceholderAPI;
 import net.kyori.adventure.text.Component;
 import org.bukkit.Bukkit;
 import org.bukkit.OfflinePlayer;
-import org.bukkit.World;
 import org.bukkit.util.CachedServerIcon;
 
 import java.awt.image.BufferedImage;
@@ -177,10 +176,7 @@ public class PaperEventWrapper implements GenericEventWrapper<CachedServerIcon, 
     
     @Override
     public GenericServer createGenericServer(int playersOnline, int playersMax, String host){
-        Map<String, World> worlds = new HashMap<>();
-        plugin.getServer().getWorlds().forEach(world -> worlds.put(world.getName(), world));
-        
-        return new BukkitServerImpl(worlds, playersOnline, playersMax, host);
+        return new BukkitServerImpl(plugin.getWorldCache().worlds(), playersOnline, playersMax, host);
     }
     
     @Override
