+++ velocity/src/main/java/ch/andre601/advancedserverlist/velocity/VelocityCore.java
@@ -30,8 +30,8 @@ import ch.andre601.advancedserverlist.core.interfaces.PluginLogger;
 import ch.andre601.advancedserverlist.core.interfaces.core.PluginCore;
 import ch.andre601.advancedserverlist.core.profiles.favicon.FaviconHandler;
 import ch.andre601.advancedserverlist.velocity.commands.CmdAdvancedServerList;
-import ch.andre601.advancedserverlist.velocity.events.JoinEvent;
-import ch.andre601.advancedserverlist.velocity.events.PingEvent;
+import ch.andre601.advancedserverlist.velocity.listeners.JoinEvent;
+import ch.andre601.advancedserverlist.velocity.listeners.PingEvent;
 import ch.andre601.advancedserverlist.velocity.logging.VelocityLogger;
 import ch.andre601.advancedserverlist.velocity.objects.VelocityPlayerPlaceholders;
 import ch.andre601.advancedserverlist.velocity.objects.VelocityServerPlaceholders;
