+++ velocity/src/main/java/ch/andre601/advancedserverlist/velocity/listeners/PingEvent.java
@@ -23,29 +23,25 @@
  *
  */
 
-package ch.andre601.advancedserverlist.velocity.events;
+package ch.andre601.advancedserverlist.velocity.listeners;
 
+import ch.andre601.advancedserverlist.core.events.PingEventHandler;
 import ch.andre601.advancedserverlist.velocity.VelocityCore;
+import com.velocitypowered.api.event.PostOrder;
 import com.velocitypowered.api.event.Subscribe;
-import com.velocitypowered.api.event.connection.PostLoginEvent;
-import com.velocitypowered.api.proxy.Player;
+import com.velocitypowered.api.event.proxy.ProxyPingEvent;
 
-import java.net.InetSocketAddress;
-
-public class JoinEvent{
+public class PingEvent{
     
     private final VelocityCore plugin;
     
-    public JoinEvent(VelocityCore plugin){
+    public PingEvent(VelocityCore plugin){
         this.plugin = plugin;
         plugin.getProxy().getEventManager().register(plugin, this);
     }
     
-    @Subscribe
-    public void onJoin(PostLoginEvent event){
-        InetSocketAddress address = event.getPlayer().getRemoteAddress();
-        Player player = event.getPlayer();
-        
-        plugin.getCore().getPlayerHandler().addPlayer(address.getHostString(), player.getUsername(), player.getUniqueId());
+    @Subscribe(order = PostOrder.EARLY)
+    public void onProxyPing(ProxyPingEvent event){
+        PingEventHandler.handleEvent(new VelocityEventWrapper(plugin, event));
     }
 }
