+++ velocity/src/main/java/ch/andre601/advancedserverlist/velocity/listeners/PreServerListSetEventImpl.java
@@ -23,9 +23,10 @@
  *
  */
 
-package ch.andre601.advancedserverlist.velocity.events;
+package ch.andre601.advancedserverlist.velocity.listeners;
 
 import ch.andre601.advancedserverlist.api.profiles.ProfileEntry;
+import ch.andre601.advancedserverlist.velocity.events.PreServerListSetEvent;
 
 public class PreServerListSetEventImpl extends PreServerListSetEvent{
     public PreServerListSetEventImpl(ProfileEntry entry){
